1. 重新写了BuildAdjacency函数，让其构造所有elements的adjacency。
2. 重新写了ParseMSH函数。让其能够读取所有类型的element，并且储存在mesh的m_edges, m_tris, m_tets中。

3. 重新写了BuildSurfaceTriangle函数，可以通过std::vector<triangle>和std::vector<tetrahedron>来构造表面三角形。不过目前这个函数都是gpt写的，还需要验证验证。

+++ include/Types.h
@@ -10,7 +10,7 @@
 #include <vector>
 
 struct mesh_on_cpu;
-using IndexBuffer = std::vector<uint16_t>;
+using IndexBuffer = std::vector<uint32_t>;
 using VertexId = uint16_t;
 constexpr VertexId INVALID_VERTEX_ID = 0xFFFF;
